Correct expected return type for call to max() function.

+++ testsuite/theodore-papadopoulo-1.cpp
@@ -10,7 +10,7 @@ int main()
     F = -1.0, -2.0, -3.0, -4.0;
 
     TinyVector<int,1> i = maxIndex(F);
-    float f = max(F);
+    double f = max(F);
 
     BZTEST(i[0] == 0);
     BZTEST(f == -1.0);
